AS-21: Rename Routes to AppRoutes
